Always sign using replay protection
Summary: Now that the fork happened, we need to use replay protected transactions all the time. Doing otherwize will just create the opportunity to lose funds.

Test Plan:
  make check
  ../qa/pull-tester/rpc-tests.py -extended

Reviewers: freetrader, CCulianu, #bitcoin_abc

Reviewed By: CCulianu, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D455

+++ src/wallet/wallet.cpp
@@ -2598,7 +2598,7 @@ bool CWallet::FundTransaction(CMutableTransaction &tx, CAmount &nFeeRet,
 
     // Copy output sizes from new transaction; they may have had the fee
     // subtracted from them.
-    for (unsigned int idx = 0; idx < tx.vout.size(); idx++) {
+    for (size_t idx = 0; idx < tx.vout.size(); idx++) {
         tx.vout[idx].nValue = wtx.tx->vout[idx].nValue;
     }
